add back -Werror, hopefully fix CI issues

+++ src/liblsquic/lsquic_parse_iquic_common.c
@@ -315,7 +315,7 @@ static int
 popcount (unsigned v)
 {
     int count, i;
-    for (i = 0, count = 0; i < sizeof(v) * 8; ++i)
+    for (i = 0, count = 0; i < (int)sizeof(v) * 8; ++i)
         if (v & (1 << i))
             ++count;
     return count;
